Reorganize plugin files
