+++ crates/builder_/src/types/plugin/installer.rs
@@ -1,0 +1,16 @@
+use serde::Serialize;
+
+#[derive(Default, Serialize)]
+pub enum Installer {
+    Binstall(String),
+    #[default]
+    Cargo,
+    Shell(String),
+    Plugin(String),
+}
+
+impl Installer {
+    fn install(&self) {
+        todo!()
+    }
+}
